Added member events to event registry
